public/projects/image-compare/backend/requirements.txt
# Image Compare Tool - Python Backend Dependencies
# Author: Kevin Hintermaier

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
Pillow>=10.0.0
